feat: Update cloudbuild.yaml to test with JDKs 8, 21 and 24.
This also updates the build config to use a clean JDK image and the Gradle wrapper. This will prevent the need to update the cloudbuild.yaml file when Gradle is updated, as well as make the test more closely test an actual developer environment

This also adds a publish test, essentially a full build test.